Split scorecard tests into locally generated and scene history tests.
